Update tabletop-prototype.py
updated argparser code to incorporate the subparser "tts_deckbuilder". This code is not fully tested, but has been shown to work when utilizing the script through powershell:

`py -3.10  .\tabletop-prototype.py tts_deckbuilder -i <path to image to turn into a deck for TTS>`

This code was added in response to issue #3.